feat(suite): Add refferal button + refactoring footer
